fast_api/sms_gateway/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
dnspython==2.4.2
email-validator==2.1.0.post1
fastapi==0.108.0
h11==0.14.0
idna==3.6
Jinja2==3.1.2
MarkupSafe==2.1.3
pycparser==2.21
pydantic_core==2.14.6
PyJWT==2.8.0
requests==2.31.0
sniffio==1.3.0
starlette==0.32.0.post1
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.25.0
